==> 01_data_cleaning.ipynb <==
import pandas as pd
import random
import numpy as np
import math
import seaborn as sns
from datetime import date
import datetime
from sklearn import metrics
import six
from scipy import stats
import matplotlib.pyplot as plt
%matplotlib inline
from pylab import rcParams
rcParams['figure.figsize'] = 16, 9

data = pd.read_csv('HMXPC13_DI_v2_5-14-14.csv')

data = data.drop(['roles', 'incomplete_flag' , 'registered', 'certified'], axis=1) #We are dropping 3 columns:
#registered becasue it has only value 1 and nothing else; 
#incoplete flag and roles because they have mostly Null values and are only used for administration.

data.shape

At the begining of data cleaning we had 641138 rows.

#### We are checking 4 columnes- nevents, ndays_act, nplay_video and nchapters because we want to be sure they have NaN values for the state of no interaction that we can safely replace with the value zero. 

data['nevents'].value_counts(dropna = False).sort_index()

data['ndays_act'].value_counts(dropna = False).sort_index()

data['nplay_video'].value_counts(dropna = False).sort_index()

data['nchapters'].value_counts(dropna = False).sort_index()

#### Considering that all 4 tested columns have no value 0 we will now assign 0 to all rows that have NaN value

data['nevents'].fillna(0.0, inplace = True)

data['ndays_act'].fillna(0.0, inplace = True)

data['nplay_video'].fillna(0.0, inplace = True)

data['nchapters'].fillna(0.0, inplace = True)

#### Transforming Country of origin data using get_dummies function - Converts categorical variable into dummy/indicator variables

data = data.join(pd.get_dummies(data['final_cc_cname_DI']))

#### Derive column Total Activity

When the date is missing from the column "last_event_DI" this means that there was no interaction with the course after enrolment. Considering this we have decided to put starting date as the date of last interaction instead of NaN value. This way we keep all entries and we get 0 days active.

data["last_event_DI"] = data.apply(lambda x: x.start_time_DI if pd.isnull(x.last_event_DI) else x.last_event_DI, axis =1)

We have checked the type of all dates entered in columnes "start_time_DI" and "last_event_DI" and they are all String. The code that we used is removed as it takes longer time to run.
#### We now make a derivative columne that shows "total_activity" and it is determined from start time of the course and last event
To work with dates, we need to convert them from strings into a data format built for processing dates. The pandas library comes with a Timestamp data object for storing and working with dates.

data["start_time_DI"] = pd.to_datetime(data["start_time_DI"] )

data["last_event_DI"] = pd.to_datetime(data["last_event_DI"] ) #transfering dates so we can substract them to get days

data['total_activity'] = data['last_event_DI'] - data['start_time_DI']

data['total_activity'] = data['total_activity'].apply(lambda x: x.days)

#data['YoB'].fillna(int(data['YoB'].mean()), inplace = True) #for missing values we set mean of all ages

#### Dealing with NaN values and converting Level of education from categorical variable into indicator variables

data['LoE_DI'] = data['LoE_DI'].fillna('UnknownLoE') #for missing values we assigned Unknown level of Education this is not
# so high precentage of data as shown in the next cell

data['LoE_DI'].value_counts(dropna= False)

data = data.join(pd.get_dummies(data['LoE_DI']))

#### Dealing with NaN values and converting Gender from categorical variable into indicator variables

data['gender'].fillna('u', inplace = True) #changing NaN values to 'u' -unknown, again this is not too high precentage of data

data['gender'].value_counts(dropna= False)

data = data.join(pd.get_dummies(data['gender']))

#### Some grades are empty string (space) and we drop those rows because we can work only with rows that have a grade as indicator of success on course

data = data.loc[data['grade'] != " ",] 

data.dropna(axis=0, inplace=True) #We drop all rows that have NaN after cleaning

#### Transforming Year of birth so we can derive column Age
Based on multiple runs of code and investigations of data we reached a conclusion that columns that have a missing value for year of birth also have missing values for most of other columns like gender and level of education. After trying different options of assigning values and checking results we can say that the best result is when we dont fill in these values but just drop all rows with missing year of birth. From remaining rows we derive column "age".

now = datetime.datetime.now()

data['age'] = data['YoB'].apply(lambda x: now.year - int(x)) # we use provided year of birth to calculate
#age of all participants and save it as new column

data['grade'] = data['grade'].astype(float)

data.columns

data['gender'].value_counts(dropna= False)

#### At the end of data cleaning we have 494790 rows which is around 77% of the begining dataset size. Because the number of rows is still significant and most of removed rows had a lot of missing information we consider this a good result.

data.to_csv('cleaned_data.csv', index = False)
