==> 02_descriptive.ipynb <==
import pandas as pd
import numpy as np
import math
import seaborn as sns
from datetime import date
import datetime
from sklearn import metrics
import sklearn as sk
import six
from scipy import stats
import matplotlib.pyplot as plt
%matplotlib inline
from pylab import rcParams
rcParams['figure.figsize'] = 16, 9

data = pd.read_csv('cleaned_data.csv')

# We get a global variable of rowCount 
rowCount = data.shape[0]
print("Total Number of Rows:",rowCount)

data.columns

def percentage(part, whole):
  return 100 * float(part)/float(whole)

# --------------------------------------------------- Viewed ---------------------------------------------------------------- # 
print("All unique values for viewed are:",data['viewed'].unique())
# [0 1] shows that either the person has viewed a course or not.
# No need to clean any data in this column as no NaN values exist.

viewed = data.query('viewed == 1')['viewed'].count()
print("Number of people who viewed the courses is:",viewed)

perViewed = percentage(viewed, rowCount)
print(perViewed,"% of people viewed the courses")

# --------------------------------------------------- Explored ---------------------------------------------------------------- # 
print("All unique values for explored are:",data['explored'].unique())
# [0 1] shows that either the person has explored a course or not.
# No need to clean any data in this column as no NaN values exist.

explored = data.query('explored == 1')['explored'].count()
print("Number of people who explored highest parts of courses:",explored)

perExplored = percentage(explored, rowCount)
print("Only",perExplored,"% people explored the highest parts of courses")
