==> 04_linear_regression_model.ipynb <==
import pandas as pd
import numpy as np
import math
import seaborn as sns
from datetime import date
import datetime
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score
from sklearn import metrics
import sklearn as sk
import six
from scipy import stats
import matplotlib.pyplot as plt
%matplotlib inline
from pylab import rcParams
rcParams['figure.figsize'] = 16, 9

data = pd.read_csv('cleaned_data.csv')

# Multicollinearity checking

#### Based on research we drop one of dummy variables for each group (Gender, Year of birth, Level of education) of dummy created variables. This aproach is based on seperating one referent variable, usually the one with most rows.

variables = data.drop(['course_id', 'userid_DI', 'grade', 'Unknown/Other', 'UnknownLoE','u', 'o', 'm', 'YoB' ], axis=1)
#droping:
# descriptive columns that are not considered as variables like course id
# grade
# u and o columns because they have no remining values
# one referent variable for each group of dummy variables

correlations = variables.corr()

w, v = np.linalg.eig(correlations) #compute the eigenvalues and right eigenvectors
# w is array of eigenvalues
# v is array of normalized eigenvectors

df = pd.DataFrame(columns= ['variables', 'w'])

df['variables'] = correlations.columns
df['w'] = w

df

We see that eigenvalues are closest to 0 for variables nevents, ndays_active, nplay_videos and nchapters. This could possibly be explained by saying that all of them measure some types of interaction with course materials so it makes sense that they are more strongly correlated. 
#### Based on research of how to deal with multicollinearity we have decided to remove two of them from final set of variables. This will solve the problem of existance of multicollinearity.

## Multicollinearity checking without dummy variables
We were curious too see the result in this special case and decided to leave also this check for final version.

colums = ['viewed', 'explored', 'nevents', 'ndays_act',
       'nplay_video', 'nchapters', 'nforum_posts', 'total_activity',
       'age']

n_c = data[colums].corr()

w, v = np.linalg.eig(n_c)

df1 = pd.DataFrame(columns= ['variables', 'w'])

df1['variables'] = n_c.columns
df1['w'] = w

df1

Lowest eigenvalue is for column total_activity.

#### The good confirmation of our analysis is the fact that correlation analysis in previous notebook and the multicollinearity analysis preforemed here both point out that variables that measure interaction with courses materials are mostly correlated with each other and that this needs to be handeld so it wouldn't affect our model.

# Linear regression

excluded = ['course_id', 'o', 'u','userid_DI', 'final_cc_cname_DI', 'start_time_DI', 'last_event_DI', 'LoE_DI','YoB',
            'gender','grade'] 

regression_data = data.drop(excluded, axis=1) # excluding all columns that are not suitable variables for linear regression

cols = ['nchapters', 'total_activity', 'nplay_video', 'age']

X, y = regression_data.drop(cols, axis=1), data['grade']
lm = LinearRegression()
lm.fit(X, y)
predictions = lm.predict(X)
print('R2:', metrics.r2_score(y, predictions))

list(zip(X.columns, lm.coef_))

lm.intercept_

sk.feature_selection.f_regression(X,y)
